Updated System/OnlineExamPortal/requirements.txt
Django==3.1.7
djangorestframework==3.13.1
mysqlclient